tests/wpt/web-platform-tests/css/css-box/parsing/clear-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: getComputedStyle().clear</title>
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-clear">
<link rel="help" href="https://drafts.csswg.org/css-logical/#float-clear">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("clear", "none");
test_computed_value("clear", "left");
test_computed_value("clear", "right");
test_computed_value("clear", "both");

test_computed_value("clear", "inline-start");
test_computed_value("clear", "inline-end");
</script>
</body>
</html>
